Move jsonStringify test from atomic/regexp.spec.ts
- Moved test from packages/run-types/src/runType/atomic/regexp.spec.ts:38-46
- Fixed linting issues (import style, duplicate test title)
- All tests passing
